Update controller danh mục - thay String thành ModelAndView cho một số endpoint

+++ FlowieeOfficial/src/main/java/com/flowiee/app/danhmuc/service/impl/LoaiTaiLieuServiceImpl.java
@@ -1,8 +1,16 @@
 package com.flowiee.app.danhmuc.service.impl;
 
+import com.flowiee.app.common.exception.BadRequestException;
+import com.flowiee.app.common.utils.IPUtil;
 import com.flowiee.app.danhmuc.entity.LoaiTaiLieu;
 import com.flowiee.app.danhmuc.repository.LoaiTaiLieuRepository;
 import com.flowiee.app.danhmuc.service.LoaiTaiLieuService;
+import com.flowiee.app.hethong.entity.Account;
+import com.flowiee.app.hethong.entity.SystemLog;
+import com.flowiee.app.hethong.model.SystemLogAction;
+import com.flowiee.app.hethong.model.action.DanhMucAction;
+import com.flowiee.app.hethong.model.module.SystemModule;
+import com.flowiee.app.hethong.service.SystemLogService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -12,6 +20,8 @@ import java.util.List;
 public class LoaiTaiLieuServiceImpl implements LoaiTaiLieuService {
     @Autowired
     private LoaiTaiLieuRepository loaiTaiLieuRepository;
+    @Autowired
+    private SystemLogService systemLogService;
 
     @Override
     public List<LoaiTaiLieu> findAll() {
@@ -24,7 +34,7 @@ public class LoaiTaiLieuServiceImpl implements LoaiTaiLieuService {
     }
 
     @Override
-    public LoaiTaiLieu findById(int id) {
+    public LoaiTaiLieu findById(Integer id) {
         return loaiTaiLieuRepository.findById(id).orElse(null);
     }
 
@@ -38,17 +48,28 @@ public class LoaiTaiLieuServiceImpl implements LoaiTaiLieuService {
         if (findByTen(loaiTaiLieu.getTen()) != null) {
             return null;
         }
-        return loaiTaiLieuRepository.save(loaiTaiLieu);
-    }
-    @Override
-    public void update(LoaiTaiLieu loaiSanPham, int id) {
-        loaiSanPham.setId(id);
-        loaiTaiLieuRepository.save(loaiSanPham);
+        LoaiTaiLieu loaiTaiLieuSaved = loaiTaiLieuRepository.save(loaiTaiLieu);
+        systemLogService.writeLog(SystemModule.KHO_TAI_LIEU.name(), SystemLogAction.THEM_MOI.name(), loaiTaiLieu.toString(), null);
+        return loaiTaiLieuSaved;
     }
 
     @Override
-    public boolean delete(int id) {
+    public void update(LoaiTaiLieu loaiTaiLieu, Integer id) {
+        if (id != null && this.findById(id) != null) {
+            loaiTaiLieu.setId(id);
+            LoaiTaiLieu loaiTaiLieuUpdated = loaiTaiLieuRepository.save(loaiTaiLieu);
+            systemLogService.writeLog(SystemModule.KHO_TAI_LIEU.name(), DanhMucAction.UPDATE_DANHMUC.name(), loaiTaiLieu.toString(), loaiTaiLieuUpdated.toString());
+        }
+    }
+
+    @Override
+    public boolean delete(Integer id) {
+        LoaiTaiLieu loaiTaiLieuToDelete = findById(id);
+        if (loaiTaiLieuToDelete == null) {
+            throw new BadRequestException();
+        }
         loaiTaiLieuRepository.deleteById(id);
+        systemLogService.writeLog(SystemModule.KHO_TAI_LIEU.name(), DanhMucAction.UPDATE_DANHMUC.name(), loaiTaiLieuToDelete.toString(), null);
         if (findById(id) == null){
             return true;
         } else {
